+++ CHANGES.md
@@ -25,6 +25,7 @@ Build 🧱:
 
 Other changes:
  - Use File extension functions to make code more concise (#1996)
+ - Create a script to import SAS strings (#1909)
 
 Changes in Element 1.0.5 (2020-08-21)
 ===================================================